+++ R/wham_plots_tables.R
@@ -13,12 +13,13 @@ plot.ecov <- function(mod, plot.pad = FALSE, do.tex=FALSE, do.png=FALSE, fontfam
     sdrep = mod$sdrep
   }
 
-  if("Ecov_obs_logsigma" %in% names(mod$env$par)){
-    ecov.obs.sig = matrix(exp(sdrep[rownames(sdrep) %in% "Ecov_obs_logsigma",1]), ncol=dat$n_Ecov) # all the same bc obs_sig_var --> 0
-    if(dim(ecov.obs.sig)[1] == 1) ecov.obs.sig = matrix(rep(ecov.obs.sig, dim(ecov.obs)[1]), ncol=dat$n_Ecov, byrow=T)
-  } else {
-    ecov.obs.sig = exp(mod$input$par$Ecov_obs_logsigma)
-  }
+  ecov.obs.sig = mod$rep$Ecov_obs_sigma
+  # if("Ecov_obs_logsigma" %in% names(mod$env$par)){
+  #   ecov.obs.sig = matrix(exp(sdrep[rownames(sdrep) %in% "Ecov_obs_logsigma",1]), ncol=dat$n_Ecov) # all the same bc obs_sig_var --> 0
+  #   if(dim(ecov.obs.sig)[1] == 1) ecov.obs.sig = matrix(rep(ecov.obs.sig, dim(ecov.obs)[1]), ncol=dat$n_Ecov, byrow=T)
+  # } else {
+  #   ecov.obs.sig = exp(mod$input$par$Ecov_obs_logsigma)
+  # }
   ecov.use = dat$Ecov_use_obs[1:dat$n_years_Ecov,,drop=F]
   ecov.obs.sig = ecov.obs.sig[1:dat$n_years_Ecov,,drop=F]
   ecov.obs.sig[ecov.use == 0] <- NA
@@ -51,6 +52,28 @@ plot.ecov <- function(mod, plot.pad = FALSE, do.tex=FALSE, do.png=FALSE, fontfam
     points(years, ecov.obs[,i], pch=19)
     lines(years_full, ecov.pred[,i], col=plot.colors[i], lwd=3)
     title (paste0("Ecov ",i, ": ",unlist(dat$Ecov_label)[i]), outer=T, line=-1)
+    if(dat$n_years_proj_Ecov > 0) abline(v=tail(years,1), lty=2)
+
+    if(do.tex | do.png) dev.off() else par(origpar)
+  }
+  #plot standard deviations if estimated as random effects
+  ecov.obs.sigma.cv = as.list(mod$sdrep, "Std")$Ecov_obs_sigma_re
+  for (i in ecovs) if(dat$Ecov_obs_sigma_opt[i]==4)
+  {
+    if(do.tex) cairo_pdf(file.path(od, paste0("Ecov_",i,"_sig_re.pdf")), family = fontfam, height = 10, width = 10)
+    if(do.png) png(filename = file.path(od, paste0("Ecov_",i,'_sig_re.png')), width = 10*144, height = 10*144, res = 144, pointsize = 12, family = fontfam)
+    sigmas = ecov.obs.sig[,i]
+    sigmas.cv = ecov.obs.sigma.cv[,i]
+    sigmas.low <- exp(log(sigmas[,i]) - 1.96 * sigmas.cv[,i])
+    sigmas.high <- exp(log(sigmas[,i]) + 1.96 * sigmas.cv[,i])
+    y.min = min(sigmas.low, na.rm =T)
+    y.max = max(sigmas.high, na.rm =T)
+    plot(years_full, sigmas[,i], type='n', xlab="Year", ylab=unlist(dat$Ecov_label)[i],
+         ylim=c(y.min, y.max))
+    polygon(c(years_full,rev(years_full)), c(sigmas.low, rev(sigmas.high)), col=adjustcolor(plot.colors[i], alpha.f=0.4), border = "transparent")
+    points(years, sigmas[,i], pch=19)
+    lines(years_full, sigmas[,i], col=plot.colors[i], lwd=3)
+    title (paste0("Ecov ",i, ": ",unlist(dat$Ecov_label)[i], " SD"), outer=T, line=-1)
     if(dat$n_years_proj_Ecov > 0) abline(v=tail(years,1), lty=2)
 
     if(do.tex | do.png) dev.off() else par(origpar)